data type fix and user delete

+++ src/main/java/com/tesis/controllers/UserController.java
@@ -36,13 +36,13 @@ public class UserController {
     public Object getUsersByUserID(Request request, Response response) throws ApiException {
 
         String param = request.params("user_id");
-        Integer userID;
+        Long userID;
         if (StringUtils.isBlank(param)) {
             throw new ApiException("invalid_data", "[reason: invalid_user_id] [method: UserController.getUserByUserID]");
         }
 
         try {
-            userID = Integer.valueOf(param);
+            userID = Long.valueOf(param);
         } catch (NumberFormatException e) {
             throw new ApiException("invalid_data", "[reason: invalid_user_id] [method: UserController.getUserByUserID]");
         }
@@ -73,13 +73,13 @@ public class UserController {
     public  Object updateUser(Request request, Response response) throws ApiException {
 
         String param = request.params("user_id");
-        Integer userID;
+        Long userID;
         if (StringUtils.isBlank(param)) {
             throw new ApiException("invalid_data", "[reason: invalid_user_id] [method: UserController.updateUser]");
         }
 
         try {
-            userID = Integer.valueOf(param);
+            userID = Long.valueOf(param);
         } catch (NumberFormatException e) {
             throw new ApiException("invalid_data", "[reason: invalid_user_id] [method: UserController.updateUser]");
         }
@@ -94,12 +94,12 @@ public class UserController {
 
     public Object deleteUser(Request request, Response response) throws  ApiException {
         String param = request.params("user_id");
-        Integer userID;
+        Long userID;
         if (StringUtils.isBlank(param)) {
             throw new ApiException("invalid_data", "[reason: invalid_user_id] [method: UserController.deleteUser]");
         }
         try {
-            userID = Integer.valueOf(param);
+            userID = Long.valueOf(param);
         } catch (NumberFormatException e) {
             throw new ApiException("invalid_data", "[reason: invalid_user_id] [method: UserController.deleteUser]");
         }
