+++ src/main/java/com/tesis/jooq/tables/daos/DevicesDao.java
@@ -82,7 +82,7 @@ public class DevicesDao extends DAOImpl<DevicesRecord, com.tesis.jooq.tables.poj
     /**
      * Fetch records that have <code>user_id IN (values)</code>
      */
-    public List<com.tesis.jooq.tables.pojos.Devices> fetchByUserId(Integer... values) {
+    public List<com.tesis.jooq.tables.pojos.Devices> fetchByUserId(Long... values) {
         return fetch(Devices.DEVICES.USER_ID, values);
     }
